replace prometheus with influxdb (mostly)
everything piped together ok in this iteration
but dashboard queries effectively need to be re-written to use flux (sql like influx query language)

enum impl would probably need to be done in code now instead of delegating it to the prom_client/promdb

basically need something so when we `publish_enum_metric` for some value, it does the work to set the value of that enum to 1 and all other enum values to 0

or we refactor the publish_enum behavior to be a tag we put onto other metrics but that would be a more wide reaching change.

alternatively could store the enum value as a string, though this may not play nice with the state timeline component on the dashboard. if this works it is probably lowest effort for future maintenance though