refactor avionics vehicle telem reads to use the stream API
only migrated 1 scenario in this iteration and focused only on GNC mode 1-3, where many overruns were observed previously

this is mostly a POC to measure impact, should proably merge this with the Telemetry and/or KSPTelemetry class for an operationalized solution

the impact of this is nearly 0 overruns at 50hz, with exception of ~20% overruns in mode 2a (potentially due to linear algebra or lots of RPC set calls). at 40hz there are basically no overruns after this change.

see also https://krpc.github.io/krpc/python/client.html#streaming-data-from-the-server